Attempt underwater physics & facingDir
it's very buggy rn. at least the multiplier portion of underwater physics seems pretty okay. but facingDir is a problem. for some reason it seems like the two hair segments on front of the head go up while also going up, esp. underwater. what's up with that? it might have to do with how  the gravity is set to make the hair point up when underwater. will test later.